% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.get.R
\name{storage.get}
\alias{storage.get}
\title{Get data object from storage}
\usage{
storage.get(storage.name, id)
}
\arguments{
\item{storage.name}{storage name}

\item{id}{Object id}
}
\value{
data from Bigstream storage
}
\description{
read data from Bigstream storage via Bigstream storage API
}
\examples{
host <- "http://sample.bigstream.io"
port <- 19080
storage_name <- "sample.sensordata"
token <- "token"
conn <- bs.connect(host, port, token)
storage.get(storage_name,id="010000000000000258107ea7")
}
