% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.getJSON.R
\name{storage.getJSON}
\alias{storage.getJSON}
\title{Get data object from storage}
\usage{
storage.getJSON(storage_name, id = NULL, index = NULL, key = NULL)
}
\arguments{
\item{storage_name}{storage name}

\item{id}{Object id}

\item{index}{get data by index}

\item{key}{get data by defined key  in storage}
}
\value{
data in JSON format from Bigstream storage. Return lastest data if no id, index or key param
}
\description{
read data from Bigstream storage via Bigstream storage API
This command support on Bigstream version 1.2 or more only
}
\examples{
host <- "http://sample.bigstream.io"
storage_name <- "sample.sensordata"
token <- "token"
bs.connect(host, token)

#return lastest data
storage.getJSON(storage_name)

# return data object 010000000000000258107ea7
storage.getJSON(storage_name,id="010000000000000258107ea7")

# return data index 1 (1st data in storage)
storage.getJSON(storage_name,index=1)

# return data by key defined
storage.getJSON(storage_name,key="")
}
