% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.list.R
\name{storage.list}
\alias{storage.list}
\title{Show storage list}
\usage{
storage.list()
}
\value{
Storage list in give Bigstream connection
}
\description{
Given storage list from current connection setup or given connection
}
\examples{
host <- "http://sample.bigstream.io"
storage_name <- "sample.sensordata"
token <- "token"
bs.connect(host, token)
storage.list()

}
