% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.put.R
\name{storage.put}
\alias{storage.put}
\title{Put data to storage}
\usage{
storage.put(storage_name, x)
}
\arguments{
\item{storage_name}{target storage to put data}

\item{x}{data frame object}
}
\value{
put status
}
\description{
Put data to storage
}
\examples{
host <- "http://sample.bigstream.io"
port <- 19080
storage_name <- "sample.sensordata"
token <- "token"
conn <- bs.connect(host, port, token)
storage.put(storage_name,x)
}
