% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.save.R
\name{storage.save}
\alias{storage.save}
\title{Save data to storage}
\usage{
storage.save(storage_name, x)
}
\arguments{
\item{storage_name}{target storage to put data}

\item{x}{data frame object}
}
\value{
save status
}
\description{
This function replace data on given Bigstream Storage name.
}
\examples{
host <- "http://sample.bigstream.io"
port <- 19080
storage_name <- "sample.sensordata"
token <- "token"
conn <- bs.connect(host, port, token)
storage.save(storage_name,x)
}
