% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.stat.R
\name{storage.stat}
\alias{storage.stat}
\title{Get storage information}
\usage{
storage.stat(storage_name)
}
\arguments{
\item{storage_name}{storage name}
}
\value{
Storage information
}
\description{
Get storage information
}
\examples{
host <- "http://sample.bigstream.io"
storage_name <- "sample.sensordata"
token <- "token"
bs.connect(host, storage_name, token)
storage.stat(storage_name)
}
