% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.connect.R
\name{bs.connect}
\alias{bs.connect}
\title{Setup Bigstream connector function}
\usage{
bs.connect(host, token)
}
\arguments{
\item{host}{Bigstream Service host.}

\item{token}{Set API token}
}
\value{
connection setting
}
\description{
Setup Bigstream connector function
}
\examples{
host <- "http://sample.bigstream.io/"
token <- "token"
bs.connect(host, storage_name, token)
}
