% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.R
\name{storage.list}
\alias{storage.list}
\title{Show storage list}
\usage{
storage.list(conn = NULL)
}
\arguments{
\item{conn}{Bigstream connection (No need if already call bs.connect() )}
}
\value{
Storage list in give Bigstream connection
}
\description{
Given storage list from current connection setup or given connection
}
\examples{
host <- "http://sample.bigstream.io"
port <- 19080
storage_name <- "sample.sensordata"
token <- "token"
conn <- bs.connect(host, port, token)
storage.list()

or

storage.list(conn)
}
