% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.storage.read.R
\name{storage.read}
\alias{storage.read}
\title{Get data object from storage}
\usage{
storage.read(storage.name, field = c("data", "id", "meta", "_data", "_id",
  "_meta"), from = 1, offset = NULL, last = 0, limit = 10,
  flatten = TRUE)
}
\arguments{
\item{storage.name}{storage name}

\item{field}{Specify object output "id","meta" or "data"}

\item{from}{start from object index default is 1}

\item{offset}{Lookup data after given object id}

\item{last}{number of lastest data to read (overwrite "from" arguments)}

\item{limit}{limit object output}

\item{flatten}{return flatten Data.frame or return JSON if option is FALSE}
}
\value{
data from Bigstream storage
}
\description{
read data from Bigstream storage via Bigstream storage API
}
\examples{
host <- "http://sample.bigstream.io"
storage_name <- "sample.sensordata"
token <- "token"
bs.connect(host, storage_name, token)
storage.read(storage_name)
}
